requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==2.1.3
Pillow==10.0.1
pdf2image==1.16.3
Werkzeug==2.3.7
PyMuPDF==1.26.5